requirements.txt
pyyaml>=6.0
requests>=2.31.0
schedule>=1.2.0